+++ components/UploadPopover/UploadPopover.module.css
@@ -6,7 +6,7 @@
   border: 2px solid rgba(234, 234, 234, 0.4);
   border-radius: 8px;
   width: 330px;
-  height: 280px;
+  height: 290px;
   padding-top: 10px;
   padding-left: 30px;
   padding-right: 30px;
