+++ pages/transactions/Transactions.module.css
@@ -92,115 +92,6 @@
   margin-top: 10px;
 }
 
-.filter-row {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  height: 70px;
-  margin-left: 2%;
-  margin-right: 2%;
-  padding-top: 1%;
-  padding-bottom: 1%;
-}
-
-.date-range {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 300px;
-}
-
-.calendar-box {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--sapphire-light);
-  height: 35px;
-  width: 35px;
-  border-radius: 6px;
-}
-
-.calendar-icon {
-  width: 22px;
-  height: 22px;
-}
-
-.date-display {
-  display: flex;
-  align-items: center;
-  background-color: var(--sapphire-light);
-  padding: 8px;
-  padding-left: 12px;
-  padding-right: 12px;
-  border-radius: 7px;
-  font-size: 12px;
-  margin-left: 5px;
-  margin-right: 10px;
-  height: 35px;
-}
-
-.chevron-wrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chevron-left {
-  fill: none;
-  cursor: pointer;
-  width: 15px;
-  height: 15px;
-}
-
-.chevron-right {
-  fill: none;
-  cursor: pointer;
-  width: 15px;
-  height: 15px;
-}
-
-.filters {
-  display: flex;
-  align-items: center;
-  width: 45%;
-}
-
-.filter-select {
-  width: 25%;
-  min-width: 80px;
-}
-
-.filter-label {
-  font-size: 12px;
-  font-weight: 800;
-}
-
-.select {
-  height: 35px;
-  align-items: center;
-  background-color: var(--sapphire-light);
-  border-radius: 12px;
-  border: 0px solid black;
-}
-
-.search-bar {
-  border: 1px solid #f0f0f0;
-  border-radius: 19.5px;
-  width: 30%;
-  min-width: 175px;
-  height: 36px;
-  font-family: Avenir;
-  font-size: 12px;
-  padding-left: 1%;
-}
-
-.search-icon {
-  margin-right: 3%;
-  width: 20px;
-  height: 20px;
-}
-
 .children {
   margin: 0px;
 }
